Changed the method of writing text to the terminal because the old solution of using std::cout for each character individually was very slow. Instead fwrite is used instead to write the front buffer in row sized chunks.
This new functionality involves first separating the entire buffer into individual char arrays that are just rows. Then the row can be written and then a newline character is often written. One potential issue with this method is that there is a large amount of dynamic memory allocation, which might hurt performance and introduce bugs. Fundamental changes to Screen would help if I could pre-insert newline characters at the appropriate places in the buffers and then prevent any drawing functions from editing those entries. Then the file API would be able to handle the buffering even more and it would likely be faster.

Another item of note is that a flush of the stdout buffer occurs every frame. So far, this seems to cause no performance issues. However, it might later be an issue.